Implement fire weather index (#226)
* added original fwi code and test

* converted fwi into individual stand-alone functions

* xarray version of fwi

* added docstrings

* updated history

* pep8

* removed unused code

* added test for gwfed. bug with array init values

* added gfwed test. fails

* added a TestData class

* using new TestData setup for test_fwi

* trying to implement gfwed algorithm

* trying big iterator to include start-stop logic

* Draft on master function for FWI

* WIP on big iterator

* WIP - Working version of a big iterator and its xr method

* nan-proofing of drought_code -

* Fix of duff_moisture for invalid values

* [WIP] Working version of FWI-computation

* Organized ufunc, calc func and two indices

* cadefactor fixes

* Rewrite tests, add doc, fix some caveats

* Expose drougt_code and update history.rst

* I always forget to add the test files

* rollback to include CFS tests and FWI TestDataSet - Tests will fail

* Edited FWI tests, add some doc, skip GFWED tests

* Remove FWI from data.py

* Move params info to module doc

+++ xclim/run_length.py
@@ -462,7 +462,7 @@ def first_run_ufunc(
 
     if index is not None and ~(np.isnan(ind)):
         val = getattr(x.indexes["time"], index)
-        i = int(ind.data)
+        i = ind.data.astype(int)
         ind.data = val[i]
 
     return ind
